refactor: rename package 'mapper' to 'convert' and update method names and references accordingly
