+++ src/main/java/com/henriques/task_manager/service/TaskService.java
@@ -1,13 +1,10 @@
 package com.henriques.task_manager.service;
 
-import com.henriques.task_manager.api.Priority;
-import com.henriques.task_manager.api.Status;
 import com.henriques.task_manager.api.TaskDTO;
 import com.henriques.task_manager.exceptions.TaskNotFoundException;
-import com.henriques.task_manager.mapper.TaskConvert;
+import com.henriques.task_manager.convert.TaskConvert;
 import com.henriques.task_manager.model.TaskEntity;
 import com.henriques.task_manager.repository.TaskRepository;
-import jakarta.annotation.PostConstruct;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
@@ -30,7 +27,7 @@ public class TaskService {
     public void saveTask(TaskDTO taskDTO) {
 
         try {
-            TaskEntity taskEntity = taskConvert.convert(taskDTO);
+            TaskEntity taskEntity = taskConvert.convertTaskDtoToTaskEntity(taskDTO);
             taskRepository.save(taskEntity);
         } catch (RuntimeException e) {
             throw new RuntimeException("Erro ao salvar a tarefa: " + e.getMessage());
@@ -41,7 +38,7 @@ public class TaskService {
         Optional<TaskEntity> optionalTask = taskRepository.findById(id);
 
         if (optionalTask.isPresent()) {
-            return taskConvert.convert(optionalTask.get());
+            return taskConvert.convertTaskEntityToTaskDto(optionalTask.get());
         } else {
             throw new TaskNotFoundException("Task with ID " + id + " not found.");
         }
@@ -77,7 +74,7 @@ public class TaskService {
     public List<TaskDTO> getAllTasks() {
         return taskRepository.findAllByOrderByCreatedOnDesc()
                 .stream()
-                .map(taskConvert::convert)
+                .map(taskConvert::convertTaskEntityToTaskDto)
                 .collect(Collectors.toList());
     }
 }
